speed up lsp-position-to-point under eglot-imenu
The whole Imenu alist is being computed at once, so nothing can change
during that.

For the position translator, specifically, the implication is that it
can punt of saving excursion/restriction, assume that the buffer is
widened, and perform relative line movements instead of traversing the
buffer from the beginning each time.